backend/requirements.txt
requests
psycopg2-binary
python-dotenv
flask
flask-cors
